common/requirements.txt
hvac
fastapi[all]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pyyaml
redis